refactor : HttpRequest
-접근 제어자 변경
-메소드명 변경
-패키지명 변경
